backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
scikit-learn>=1.2.2
requests>=2.28.0
beautifulsoup4>=4.12.0
arxiv>=1.4.7
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
pymongo>=4.3.0
apscheduler>=3.10.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5
alembic>=1.10.0
asyncpg>=0.27.0
openai>=1.0.0
pymupdf>=1.22.0 